added User-Commands and Message-Commands
