+++ src/main/java/net/dv8tion/jda/api/interactions/commands/UserCommand.java
@@ -1,0 +1,27 @@
+package net.dv8tion.jda.api.interactions.commands;
+
+import net.dv8tion.jda.api.entities.Guild;
+import net.dv8tion.jda.api.utils.data.DataObject;
+import net.dv8tion.jda.internal.JDAImpl;
+
+/**
+ * Represents a Discord user-command.
+ * <br>This can be used to edit or delete the command.
+ *
+ * @see Guild#retrieveCommandById(String)
+ * @see Guild#retrieveCommands()
+ */
+public class UserCommand extends Command
+{
+
+    public UserCommand(JDAImpl api, Guild guild, DataObject json)
+    {
+        super(api, guild, json);
+    }
+
+    @Override
+    public CommandType getType()
+    {
+        return CommandType.USER;
+    }
+}
