+++ src/main/java/net/dv8tion/jda/api/interactions/commands/UserCommandInteraction.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2015 Austin Keener, Michael Ritter, Florian Spieß, and the JDA contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.dv8tion.jda.api.interactions.commands;
+
+import net.dv8tion.jda.api.entities.Member;
+import net.dv8tion.jda.api.entities.User;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+/**
+ * Interaction of a User-Command.
+ *
+ * @see net.dv8tion.jda.api.events.interaction.UserCommandEvent
+ */
+public interface UserCommandInteraction extends CommandInteraction
+{
+
+    /**
+     * The user that was selected
+     *
+     * @return The targeted user
+     */
+    @Nonnull
+    User getTargetUser();
+
+    /**
+     * The user that was selected
+     *
+     * @return The targeted user
+     */
+    @Nullable
+    Member getTargetMember();
+}
