+++ src/main/java/net/dv8tion/jda/internal/handle/InteractionCreateHandler.java
@@ -16,18 +16,13 @@
 
 package net.dv8tion.jda.internal.handle;
 
-import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
-import net.dv8tion.jda.api.events.interaction.GenericInteractionCreateEvent;
-import net.dv8tion.jda.api.events.interaction.SelectionMenuEvent;
-import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
+import net.dv8tion.jda.api.events.interaction.*;
 import net.dv8tion.jda.api.interactions.InteractionType;
+import net.dv8tion.jda.api.interactions.commands.CommandType;
 import net.dv8tion.jda.api.interactions.components.Component;
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.internal.JDAImpl;
-import net.dv8tion.jda.internal.interactions.ButtonInteractionImpl;
-import net.dv8tion.jda.internal.interactions.CommandInteractionImpl;
-import net.dv8tion.jda.internal.interactions.InteractionImpl;
-import net.dv8tion.jda.internal.interactions.SelectionMenuInteractionImpl;
+import net.dv8tion.jda.internal.interactions.*;
 import net.dv8tion.jda.internal.requests.WebSocketClient;
 
 public class InteractionCreateHandler extends SocketHandler
@@ -72,9 +67,24 @@ public class InteractionCreateHandler extends SocketHandler
 
     private void handleCommand(DataObject content)
     {
-        api.handleEvent(
-            new SlashCommandEvent(api, responseNumber,
-                new CommandInteractionImpl(api, content)));
+        switch (CommandType.fromKey(content.getObject("data").getInt("type")))
+        {
+        case CHAT_INPUT:
+            api.handleEvent(
+                    new SlashCommandEvent(api, responseNumber,
+                            new SlashCommandInteractionImpl(api, content)));
+            break;
+        case USER:
+            api.handleEvent(
+                    new UserCommandEvent(api, responseNumber,
+                            new UserCommandInteractionImpl(api, content)));
+            break;
+        case MESSAGE:
+            api.handleEvent(
+                    new MessageCommandEvent(api, responseNumber,
+                            new MessageCommandInteractionImpl(api, content)));
+            break;
+        }
     }
 
     private void handleAction(DataObject content)
